requirements.txt
streamlit>=1.22
openai>=1.0.0
yfinance>=0.2.28
python-dotenv>=1.0.0
pandas>=1.5
